refactor(extension): split the webview provider into modules
Follows the official webview-view-sample layout: HTML/CSS/JS move out of inline template strings into media/, and the provider, generator logic, and nonce helper each get their own module instead of one 300-line file.

While touching this, catalog data (labels, descriptions, feature IDs) now gets escaped before landing in the HTML. features.json is  local and self-authored, so this isn't urgent, but escaping costs nothing and closes the self-XSS gap for good.
