+++ extension/package.json
@@ -52,12 +52,15 @@
     "watch": "tsc -watch -p ./"
   },
   "devDependencies": {
+    "@types/escape-html": "^1.0.4",
     "@types/node": "^24.0.0",
     "@types/vscode": "^1.85.0",
+    "@types/vscode-webview": "^1.57.5",
     "@vscode/vsce": "^3.9.2",
     "typescript": "^5.6.0"
   },
   "dependencies": {
+    "escape-html": "^1.0.3",
     "jsonc-parser": "^3.3.1"
   }
 }
